refactor: rename user service package to application
