+++ src/main/java/com/handwoong/everyonewaiter/user/application/port/UserLoginService.java
@@ -1,4 +1,4 @@
-package com.handwoong.everyonewaiter.user.service.port;
+package com.handwoong.everyonewaiter.user.application.port;
 
 import com.handwoong.everyonewaiter.common.infrastructure.jwt.JwtToken;
 import com.handwoong.everyonewaiter.user.dto.UserLogin;
