+++ src/main/java/com/handwoong/everyonewaiter/user/infrastructure/UserLoginServiceImpl.java
@@ -3,9 +3,9 @@ package com.handwoong.everyonewaiter.user.infrastructure;
 import com.handwoong.everyonewaiter.common.infrastructure.jwt.JwtToken;
 import com.handwoong.everyonewaiter.common.infrastructure.jwt.JwtTokenProvider;
 import com.handwoong.everyonewaiter.common.infrastructure.jwt.TokenInfo;
+import com.handwoong.everyonewaiter.user.application.port.UserLoginService;
 import com.handwoong.everyonewaiter.user.dto.UserLogin;
 import com.handwoong.everyonewaiter.user.exception.UnauthorizedAccessException;
-import com.handwoong.everyonewaiter.user.service.port.UserLoginService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
