+++ src/test/java/com/handwoong/everyonewaiter/user/mock/FakeUserRepository.java
@@ -1,10 +1,10 @@
 package com.handwoong.everyonewaiter.user.mock;
 
+import com.handwoong.everyonewaiter.user.application.port.UserRepository;
 import com.handwoong.everyonewaiter.user.domain.User;
 import com.handwoong.everyonewaiter.user.domain.UserId;
 import com.handwoong.everyonewaiter.user.domain.Username;
 import com.handwoong.everyonewaiter.user.exception.UserNotFoundException;
-import com.handwoong.everyonewaiter.user.service.port.UserRepository;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
